Move testing to DataTest.
